+++ src/sass/styles.scss
@@ -74,7 +74,6 @@ $border-thin: 1px solid $color-mid-light !default;
 @import "cluster_list";
 @import "configuration_table";
 @import "custom_isos";
-@import "custom_select";
 @import "detail_page";
 @import "detail_panels";
 @import "disk_device_form";
